refactor display of table: create TableModel

+++ src/main/java/ijt/table/columns/IntegerColumn.java
@@ -86,7 +86,7 @@ public class IntegerColumn extends AbstractColumn implements NumericColumn
     @Override
     public Object get(int row)
     {
-        return Double.valueOf(data[row]);
+        return Integer.valueOf(data[row]);
     }
 
     @Override
